requirements.txt
# Auto News Pipeline Dependencies
feedparser>=6.0.0
groq>=0.4.0
requests>=2.31.0
python-dotenv>=1.0.0
jinja2>=3.1.0
python-slugify>=8.0.0
